Added Erlang and Python versions of the program.
The Erlang version has a bit of weirdness with regards to the height
function, and how the loop for the circle_pi implementation is handled.

It also seems to come out with different values once the numbers get
very small, perhaps it has some sort of accurate rounding implementation
that python/C don't use by default? Or it might just be a mistake.